@@ -11,13 +11,17 @@
     },
     "devDependencies": {
         "@tailwindcss/forms": "^0.2.1",
+        "@vue/compiler-sfc": "^3.0.11",
         "alpinejs": "^2.7.3",
         "autoprefixer": "^10.1.0",
         "axios": "^0.21",
         "bootstrap": "^4.6.0",
+        "css-loader": "^5.2.6",
+        "file-loader": "^6.2.0",
         "jquery": "^3.6",
         "laravel-mix": "^6.0.6",
         "lodash": "^4.17.19",
+        "mini-css-extract-plugin": "^1.6.0",
         "popper.js": "^1.16.1",
         "postcss": "^8.2.1",
         "postcss-import": "^12.0.1",
@@ -25,13 +29,17 @@
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
         "tailwindcss": "^2.0.2",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.5",
-        "vue-template-compiler": "^2.6.12"
+        "vue": "^3.0.11",
+        "vue-loader": "^16.2.0",
+        "vue-template-compiler": "^2.6.12",
+        "vue-router": "4.0.1"
     },
     "dependencies": {
+        "@vue/component-compiler-utils": "^3.2.0",
+        "chart.js": "2.9.4",
+        "element-plus": "^1.0.2-beta.45",
         "register-service-worker": "^1.7.2",
-        "vue-router": "^3.5.1",
+        "vue-lazyload": "^1.3.3",
         "vue2-transitions": "^0.3.0"
     }
 }
