@@ -1,10 +1,10 @@
 <template>
-    <div id="app">
+    <div>
         <router-view name="header"></router-view>
         <main>
-            <fade-transition origin="center" mode="out-in" :duration="250">
+            <!-- <fade-transition origin="center" mode="out-in" :duration="250"> -->
                 <router-view/>
-            </fade-transition>
+            <!-- </fade-transition> -->
         </main>
         <router-view name="footer"></router-view>
     </div>
@@ -14,6 +14,11 @@ import { FadeTransition } from "vue2-transitions";
 export default {
   components: {
     FadeTransition
+  },
+  mounted(){
+    window.vueRouteArgon = this.$router
   }
 };
+
+
 </script>
