@@ -1,81 +1,81 @@
 <template>
-    <section class="section section-shaped section-lg my-0">
-        <div class="shape shape-style-1 bg-gradient-default">
-            <span></span>
-            <span></span>
-            <span></span>
-            <span></span>
-            <span></span>
-            <span></span>
-            <span></span>
-            <span></span>
+  <div class="row justify-content-center">
+    <div class="col-lg-5 col-md-7">
+      <div class="card bg-secondary shadow border-0">
+        <div class="card-header bg-transparent pb-5">
+          <div class="text-muted text-center mt-2 mb-3">
+            <small>Sign in with</small>
+          </div>
+          <div class="btn-wrapper text-center">
+            <a href="#" class="btn btn-neutral btn-icon">
+              <span class="btn-inner--icon"
+                ><img src="/assets/img/icons/common/github.svg"
+              /></span>
+              <span class="btn-inner--text">Github</span>
+            </a>
+            <a href="#" class="btn btn-neutral btn-icon">
+              <span class="btn-inner--icon"
+                ><img src="/assets/img/icons/common/google.svg"
+              /></span>
+              <span class="btn-inner--text">Google</span>
+            </a>
+          </div>
         </div>
-        <div class="container pt-lg-md">
-            <div class="row justify-content-center">
-                <div class="col-lg-5">
-                    <card type="secondary" shadow
-                          header-classes="bg-white pb-5"
-                          body-classes="px-lg-5 py-lg-5"
-                          class="border-0">
-                        <template>
-                            <div class="text-muted text-center mb-3">
-                                <small>Sign in with</small>
-                            </div>
-                            <div class="btn-wrapper text-center">
-                                <base-button type="neutral">
-                                    <img slot="icon" src="img/icons/common/github.svg">
-                                    Github
-                                </base-button>
+        <div class="card-body px-lg-5 py-lg-5">
+          <div class="text-center text-muted mb-4">
+            <small>Or sign in with credentials</small>
+          </div>
+          <form role="form">
+            <base-input
+              formClasses="input-group-alternative mb-3"
+              placeholder="Email"
+              addon-left-icon="ni ni-email-83"
+              v-model="model.email"
+            >
+            </base-input>
 
-                                <base-button type="neutral">
-                                    <img slot="icon" src="img/icons/common/google.svg">
-                                    Google
-                                </base-button>
-                            </div>
-                        </template>
-                        <template>
-                            <div class="text-center text-muted mb-4">
-                                <small>Or sign in with credentials</small>
-                            </div>
-                            <form role="form">
-                                <base-input alternative
-                                            class="mb-3"
-                                            placeholder="Email"
-                                            addon-left-icon="ni ni-email-83">
-                                </base-input>
-                                <base-input alternative
-                                            type="password"
-                                            placeholder="Password"
-                                            addon-left-icon="ni ni-lock-circle-open">
-                                </base-input>
-                                <base-checkbox>
-                                    Remember me
-                                </base-checkbox>
-                                <div class="text-center">
-                                    <base-button type="primary" class="my-4">Sign In</base-button>
-                                </div>
-                            </form>
-                        </template>
-                    </card>
-                    <div class="row mt-3">
-                        <div class="col-6">
-                            <a href="#" class="text-light">
-                                <small>Forgot password?</small>
-                            </a>
-                        </div>
-                        <div class="col-6 text-right">
-                            <a href="#" class="text-light">
-                                <small>Create new account</small>
-                            </a>
-                        </div>
-                    </div>
-                </div>
+            <base-input
+              formClasses="input-group-alternative mb-3"
+              placeholder="Password"
+              type="password"
+              addon-left-icon="ni ni-lock-circle-open"
+              v-model="model.password"
+            >
+            </base-input>
+
+            <base-checkbox class="custom-control-alternative">
+              <span class="text-muted">Remember me</span>
+            </base-checkbox>
+            <div class="text-center">
+              <base-button type="primary" class="my-4">Sign in</base-button>
             </div>
+          </form>
         </div>
-    </section>
+      </div>
+      <div class="row mt-3">
+        <div class="col-6">
+          <a href="#" class="text-light"><small>Forgot password?</small></a>
+        </div>
+        <div class="col-6 text-right">
+          <router-link to="/register" class="text-light"
+            ><small>Create new account</small></router-link
+          >
+        </div>
+      </div>
+    </div>
+  </div>
 </template>
 <script>
-export default {};
+export default {
+  name: "login",
+  data() {
+    return {
+      model: {
+        email: "",
+        password: "",
+      },
+    };
+  },
+};
 </script>
-<style>
-</style>
+<style></style>
