@@ -1,3 +1,4 @@
+const { version } = require('laravel-mix');
 const mix = require('laravel-mix');
 
 /*
@@ -13,4 +14,4 @@ const mix = require('laravel-mix');
 
 mix.js('resources/js/app.js', 'public/js')
     .vue()
-    .sass('resources/sass/app.scss', 'public/css');
+    .sass('resources/sass/app.scss', 'public/css').version();
